requirements.txt
pillow-simd